Venky: Added new boosting machines and trainers + examples in examples/boosting.
Cosmin: Moved the LR and FLDA machines and trainers from src/scanning/profile to src/machine and src/trainer.
Cosmin: Added test programs for LR and FLDA in examples/machines.
Cosmin: Improved LR to use L1 and L2 norm regularization terms.

+++ src/core/DataSet.h
@@ -24,7 +24,7 @@ namespace Torch
 		virtual void	setTarget(long, Tensor*) = 0;
 
 		/// Query the number of examples in the dataset
-		int 		getNoExamples() const { return m_n_examples; };
+		long 		getNoExamples() const { return m_n_examples; };
 
 		/// Query if the DataSet has targets
 		bool 		hasTargets() const { return m_has_targets; };
